Remove reverse DNS stuff
Imagine you have two Fritz!Boxes, of which both claim the hostname 'fritz.box'.
If we do this crappy reverse DNS stuff and the hostname later resolves to a different IP, then
it is possible, that we accidentally update a different box.

Instead we should let the user decide, whether to connect with hostname or IP.